feat(roles): Add a schema(convension) to get default roles for the system users
Declaring a scheme to provide a convention to get the list of the roles to be applied to every new user as its default roles is a requirement.
In addition, there should be a consideration to categorize roles in two main categories:
	- system roles
	- user roles
System roles are predefined roles which are existed by default, but on the other hand, user roles are the ones created by the users themselves such as Admin or other normal users.
This commit contains classes that provide some facilities to solve the mentioned issues.
Be careful that in this design, all the user-defined roles are categorized as a UKNOWN system role with an extra field to indicate the user-defined role itself.

Refs #3